requirements.txt
uni-curses==2.1.3
pyYAML==6.0